+++ packages/terra-data-grid/src/WorklistDataGrid.jsx
@@ -173,6 +173,9 @@ function WorklistDataGrid(props) {
   const [rowSelectionModeAriaLiveMessage, setRowSelectionModeAriaLiveMessage] = useState(null);
   const inShiftUpDownMode = useRef(false);
   const multiSelectRange = useRef({ start: null, end: null });
+  const gridFocusedWhenRowSelectionDisabled = useRef(false);
+  const dataGridFuncRef = useRef();
+  const gridReceivedFocus = useRef(false);
 
   // -------------------------------------
   // useEffect Hooks
@@ -192,6 +195,12 @@ function WorklistDataGrid(props) {
     // Since the row selection mode has changed, the row selection mode needs to be updated.
     setRowSelectionModeAriaLiveMessage(intl.formatMessage({ id: hasSelectableRows ? 'Terra.worklistDataGrid.row-selection-mode-enabled' : 'Terra.worklistDataGrid.row-selection-mode-disabled' }));
 
+    if (gridReceivedFocus.current) {
+      let newFocusCell = dataGridFuncRef.current.getFocusedCell();
+      newFocusCell = { row: newFocusCell.row, col: Math.max(newFocusCell.col + (hasSelectableRows ? 1 : -1), 0) };
+      dataGridFuncRef.current.setFocusedRowCol(newFocusCell.row, newFocusCell.col, gridFocusedWhenRowSelectionDisabled.current);
+    }
+    gridFocusedWhenRowSelectionDisabled.current = false;
   // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [hasSelectableRows]);
 
@@ -244,6 +253,9 @@ function WorklistDataGrid(props) {
           onClearSelectedRows();
         }
       } else if (onDisableSelectableRows) {
+        // In order to set focus to the right cell after row selection is turned off, keep
+        // track of cell in grid that had focus when the call was initiated.
+        gridFocusedWhenRowSelectionDisabled.current = true;
         onDisableSelectableRows();
       }
     }
@@ -392,9 +404,18 @@ function WorklistDataGrid(props) {
     }
   };
 
+  const onFocus = () => {
+    gridReceivedFocus.current = true;
+  };
+
   // -------------------------------------
   return (
-    <div onKeyDown={handleKeyDown} onKeyUp={handleKeyUp} className={cx('worklist-data-grid-container')}>
+    <div
+      onKeyDown={handleKeyDown}
+      onKeyUp={handleKeyUp}
+      className={cx('worklist-data-grid-container')}
+      onFocus={onFocus}
+    >
       <DataGrid
         id={id}
         ariaLabel={ariaLabel}
@@ -412,6 +433,7 @@ function WorklistDataGrid(props) {
         onClearSelection={handleClearSelection}
         onRangeSelection={onRangeSelection}
         hasSelectableRows={hasSelectableRows}
+        ref={dataGridFuncRef}
       />
       <VisuallyHiddenText aria-live="polite" text={rowSelectionModeAriaLiveMessage} />
       <VisuallyHiddenText aria-live="polite" text={rowSelectionAriaLiveMessage} />